✨ Setup Postgres DB
Setting up a Postgres DB to be used with docker compose locally. Setup
of the schema and sample data is done using the initdb for simplicity,
rather than introducing a separate migration tool. This is sufficient
for the scope of this project, where it serves as a demonstration /
teaching tool.