🔭 Add Tracing Prep Code
Preparing tracing code to be included in the application as part of the
session. Given the limited time, "working" code doing what is required
is included, with the key hooks for that to be *used* being commented
out. Those hooks will be uncommented during the session to enable
tracing, and to lead into "what is that doing?"

This also means that all the required dependencies will be fetched
initially (via `go mod download` or otherwise) rather than increasing
at each step. This should allow for a largely "offline" session if/when
the prep has been done